add ILogger and ILogExport interface, to give export correct return type
also move logging related types to own file
