+++ src/logging/utils.ts
@@ -2,7 +2,7 @@
 // if you know you always have a log item, better to use the methods on the log item than these utility functions.
 
 import {Instance as NullLoggerInstance} from "./NullLogger";
-import type {FilterCreator, ILogItem, LabelOrValues, LogCallback} from "./LogItem";
+import type {FilterCreator, ILogItem, LabelOrValues, LogCallback} from "./types";
 import {LogLevel} from "./LogFilter";
 
 export function wrapOrRunNullLogger<T>(logItem: ILogItem | undefined, labelOrValues: LabelOrValues, callback: LogCallback<T>, logLevel?: LogLevel, filterCreator?: FilterCreator): T | Promise<T> {
